refactor: split the fallback merge out of the registry data file
image-fallbacks.ts is now data only; the FallbackDefinition contract and
the generic mergeFallbacks() live in merge-fallbacks.ts.
